Moved won auction function from fetchAPI to ListingProfile.js
