+++ js/profile-js/services/fetchapi.js
@@ -1,7 +1,7 @@
 import { getUserProfile } from "../apiprofile.js";
 import { updateUserProfile } from "../apiprofile.js";
 import {createPost} from "../apiprofile.js";
-import {showUserListings} from "./listings_profile.js";
+import {showUserListings, displayWinningAuctions} from "./listings_profile.js";
 import{SeeWinningAuction} from "../apiprofile.js";
 
 
@@ -170,51 +170,5 @@ async function loadUserListings() {
     }
   }
   loadUserListings();
-
-  async function displayWinningAuctions() {
-    const container = document.getElementById("winning_bid");
-    container.className = "grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6"; 
-    container.innerHTML = "";
-
-    const response = await SeeWinningAuction();
-    const winningAuctions = response?.data || [];
-
-    if (winningAuctions.length === 0) {
-        container.innerHTML = "<p>You have not won any auctions.</p>";
-        return;
-    }
-
-    winningAuctions.forEach((auction) => {
-        const card = document.createElement("div");
-        card.className = "shadow-lg flex flex-col items-center aspect-[3/4] p-3";
-
-        const image = document.createElement("img");
-        image.src = auction.media?.[0]?.url || "https://placehold.co/400x300?text=No+Image";
-        image.alt = auction.title;
-        image.className = "w-full h-2/3 object-cover"
-
-        const title = document.createElement("p");
-        title.textContent = auction.title;
-        title.className = "text-center font-body text-xl font-regular";
-
-        const endInfo = document.createElement("p");
-        endInfo.textContent = `Ended ${new Date(auction.endsAt).toLocaleString()}`;
-        endInfo.className = "text-sm text-gray-600";
-
-        const winningBid = document.createElement("p");
-        winningBid.textContent = "Winning bid";
-        winningBid.className = "text-sm text-green-600";
-
-        const infoContainer = document.createElement("div");
-        infoContainer.className = "flex-grow flex flex-col justify-center items-center space-y-1";
-        infoContainer.appendChild(title);
-        infoContainer.appendChild(winningBid);
-        infoContainer.appendChild(endInfo);
-
-        card.appendChild(image);
-        card.appendChild(infoContainer);
-        container.appendChild(card);
-    });
-}
 
 displayWinningAuctions();
